# src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API REST - PROG3 TD Final
    - **A** : Ouverture d'une nouvelle collectivite
    - **B-2** : Admission d'un nouveau membre (conditions renforcees)
    - **J** : Attribution d'un numero et d'un nom unique a une collectivite
    - **C** : Cotisations et paiements
    - **D** : Situation de tresorerie (comptes)
  title: Federation de Collectivites Agricoles API
  version: 0.0.3
servers:
- description: Serveur local
  url: http://localhost:8080
...
paths:
  /collectivities:
    get:
      operationId: getAllCollectivites
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CollectiviteResponse'
                type: array
          description: Liste des collectivites
      summary: Lister toutes les collectivites
      tags:
      - Collectivites
      x-accepts:
      - application/json
    post:
      description: |
        Conditions obligatoires :
        - Au moins 10 membres
        - Dont 5 avec >= 6 mois d'anciennete dans la federation
        - President, president adjoint, tresorier et secretaire obligatoires
        - Numero et nom attribues separement via PATCH /collectivities/{id}/identity
      operationId: creerCollectivite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectiviteRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectiviteResponse'
          description: Collectivite creee
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Regle metier non respectee
      summary: A - Creer une nouvelle collectivite
      tags:
      - Collectivites
      x-content-type: application/json
      x-accepts:
      - application/json
  /collectivities/{id}:
    get:
      operationId: getCollectivite
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectiviteResponse'
          description: Collectivite trouvee
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Recuperer une collectivite par ID
      tags:
      - Collectivites
      x-accepts:
      - application/json
  /collectivities/{id}/identity:
    patch:
      description: |
        Une fois le numero ou le nom attribue, il ne peut plus etre modifie.
        Toute tentative de modification retourne 409 Conflict.
      operationId: assignerIdentite
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignIdentityRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectiviteResponse'
          description: Identite attribuee
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: J - Attribuer numero et nom unique a une collectivite
      tags:
      - Collectivites
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /collectivities/{collectiviteId}/comptes/caisse:
    post:
      operationId: creerCaisseCollectivite
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaisseRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Caisse creee
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer la caisse d'une collectivite (une seule autorisee)
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
  /collectivities/{collectiviteId}/comptes/bancaire:
    post:
      operationId: creerCompteBancaireCollectivite
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompteBancaireRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Compte bancaire cree
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer un compte bancaire pour une collectivite
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
  /collectivities/{collectiviteId}/comptes/mobile-money:
    post:
      operationId: creerCompteMobileMoneyCollectivite
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompteMobileMoneyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Compte mobile money cree
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer un compte mobile money pour une collectivite
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
  /collectivities/{collectiviteId}/comptes:
    get:
      operationId: getComptesCollectivite
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CompteResponse'
                type: array
          description: Liste des comptes
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Lister les comptes d'une collectivite
      tags:
      - Comptes
      x-accepts:
      - application/json
  /federation/comptes/caisse:
    post:
      operationId: creerCaisseFederation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaisseRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Caisse creee
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer la caisse de la federation (une seule autorisee)
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
  /federation/comptes/bancaire:
    post:
      operationId: creerCompteBancaireFederation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompteBancaireRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Compte bancaire cree
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer un compte bancaire pour la federation
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
  /federation/comptes/mobile-money:
    post:
      operationId: creerCompteMobileMoneyFederation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompteMobileMoneyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Compte mobile money cree
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: D - Creer un compte mobile money pour la federation
      tags:
      - Comptes
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /comptes/{id}:
    get:
      operationId: getCompte
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompteResponse'
          description: Compte trouve
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Recuperer un compte par ID
      tags:
      - Comptes
      x-accepts:
      - application/json
components:
...
  schemas:
    CreateCollectiviteRequest:
      example:
        cotisationAnnuelleObligatoire: 200000
        ville: Antananarivo
        specialiteAgricole: Riziculture
        membres:
        - dateAdhesionFederation: 2000-01-23
          dateNaissance: 2000-01-23
          genre: MASCULIN
          adresse: adresse
          telephone: telephone
          poste: PRESIDENT
          nom: nom
          prenom: prenom
          email: email
          metier: metier
        - dateAdhesionFederation: 2000-01-23
          dateNaissance: 2000-01-23
          genre: MASCULIN
          adresse: adresse
          telephone: telephone
          poste: PRESIDENT
          nom: nom
          prenom: prenom
          email: email
          metier: metier
      properties:
        ville:
          example: Antananarivo
          type: string
        specialiteAgricole:
          example: Riziculture
          type: string
        cotisationAnnuelleObligatoire:
          example: 200000
          format: int64
          type: integer
        membres:
          items:
            $ref: '#/components/schemas/MembreFondateurRequest'
          minItems: 10
          type: array
      required:
      - cotisationAnnuelleObligatoire
      - membres
      - specialiteAgricole
      - ville
      type: object
    MembreFondateurRequest:
      example:
        dateAdhesionFederation: 2000-01-23
        dateNaissance: 2000-01-23
        genre: MASCULIN
        adresse: adresse
        telephone: telephone
        poste: PRESIDENT
        nom: nom
        prenom: prenom
        email: email
        metier: metier
      properties:
        nom:
          type: string
        prenom:
          type: string
        dateNaissance:
          format: date
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        adresse:
          type: string
        metier:
          type: string
        telephone:
          type: string
        email:
          format: email
          type: string
        poste:
          $ref: '#/components/schemas/Poste'
        dateAdhesionFederation:
          description: Date d'adhesion dans la federation (pour valider les 6 mois)
          format: date
          type: string
      required:
      - adresse
      - dateAdhesionFederation
      - dateNaissance
      - email
      - genre
      - metier
      - nom
      - poste
      - prenom
      - telephone
      type: object
    AssignIdentityRequest:
      example:
        numero: COL-001
        nom: Collectivite Verte d'Antananarivo
      properties:
        numero:
          description: Numero unique attribue par la federation (immuable)
          example: COL-001
          type: string
        nom:
          description: Nom unique attribue par la federation (immuable)
          example: Collectivite Verte d'Antananarivo
          type: string
      required:
      - nom
      - numero
      type: object
...
    CreateCaisseRequest:
      example:
        nomTitulaire: Caisse principale
      properties:
        nomTitulaire:
          example: Caisse principale
          type: string
      required:
      - nomTitulaire
      type: object
    CreateCompteBancaireRequest:
      example:
        numeroRib: "00005000000012345678901"
        nomBanque: BRED
        nomTitulaire: nomTitulaire
      properties:
        nomTitulaire:
          type: string
        nomBanque:
          $ref: '#/components/schemas/NomBanque'
        numeroRib:
          description: "Format : BBBBBGGGGGCCCCCCCCCCCКК (23 chiffres)"
          example: "00005000000012345678901"
          pattern: "^\\d{23}$"
          type: string
      required:
      - nomBanque
      - nomTitulaire
      - numeroRib
      type: object
    CreateCompteMobileMoneyRequest:
      example:
        serviceMobileMoney: ORANGE_MONEY
        nomTitulaire: nomTitulaire
        numeroTelephone: 034 00 000 00
      properties:
        nomTitulaire:
          type: string
        serviceMobileMoney:
          $ref: '#/components/schemas/ServiceMobileMoney'
        numeroTelephone:
          example: 034 00 000 00
          type: string
      required:
      - nomTitulaire
      - numeroTelephone
      - serviceMobileMoney
      type: object
    CollectiviteResponse:
      example:
        cotisationAnnuelleObligatoire: 6
        ville: ville
        dateCreation: 2000-01-23
        specialiteAgricole: specialiteAgricole
        numero: numero
        id: 0
        nombreMembresActifs: 1
        nom: nom
      properties:
        id:
          format: int64
          type: integer
        numero:
          description: Null jusqu'a l'attribution via PATCH /identity
          nullable: true
          type: string
        nom:
          nullable: true
          type: string
        ville:
          type: string
        specialiteAgricole:
          type: string
        dateCreation:
          format: date
          type: string
        cotisationAnnuelleObligatoire:
          format: int64
          type: integer
        nombreMembresActifs:
          format: int64
          type: integer
      type: object
...
    CompteResponse:
      example:
        serviceMobileMoney: ORANGE_MONEY
        numeroRib: numeroRib
        appartientFederation: true
        nomBanque: BRED
        typeCompte: CAISSE
        nomTitulaire: nomTitulaire
        solde: 6
        id: 0
        dateSolde: 2000-01-23
        collectiviteId: 1
        numeroTelephone: numeroTelephone
      properties:
        id:
          format: int64
          type: integer
        typeCompte:
          $ref: '#/components/schemas/TypeCompte'
        nomTitulaire:
          type: string
        solde:
          description: Solde en MGA
          format: int64
          type: integer
        dateSolde:
          format: date
          type: string
        collectiviteId:
          format: int64
          nullable: true
          type: integer
        appartientFederation:
          type: boolean
        nomBanque:
          $ref: '#/components/schemas/NomBanque'
        numeroRib:
          nullable: true
          type: string
        serviceMobileMoney:
          $ref: '#/components/schemas/ServiceMobileMoney'
        numeroTelephone:
          nullable: true
          type: string
      type: object
    ErrorResponse:
      example:
        message: message
        timestamp: 2000-01-23T04:56:07.000+00:00
        status: 0
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          type: integer
        message:
          type: string
      type: object
    ValidationErrorResponse:
      example:
        errors:
        - errors
        - errors
        timestamp: 2000-01-23T04:56:07.000+00:00
        status: 400
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          example: 400
          type: integer
        errors:
          items:
            type: string
          type: array
      type: object
    Genre:
      enum:
      - MASCULIN
      - FEMININ
      type: string
    Poste:
      enum:
      - PRESIDENT
      - PRESIDENT_ADJOINT
      - TRESORIER
      - SECRETAIRE
      - MEMBRE_CONFIRME
      - MEMBRE_JUNIOR
      type: string
...
    TypeCompte:
      enum:
      - CAISSE
      - BANCAIRE
      - MOBILE_MONEY
      type: string
    NomBanque:
      enum:
      - BRED
      - MCB
      - BMOI
      - BOA
      - BGFI
      - AFG
      - ACCES_BANQUE
      - BAOBAB
      - SIPEM
      type: string
    ServiceMobileMoney:
      enum:
      - ORANGE_MONEY
      - MVOLA
      - AIRTEL_MONEY
      type: string